Added test project
Also updated readme for helpers
