+++ hbs-test/hbs-test.html
@@ -1,0 +1,16 @@
+<!doctype html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<title>Handlebars + Require Test</title>
+
+</head>
+<body>
+	<h3>Here is random html page...</h3>
+	<hr />
+
+	<div id="tiny-module"></div>
+
+	<script type="text/javascript" data-main="js/requirejs.conf" src="js/require.js"></script>
+</body>
+</html>
