+++ hbs-test/js/handlebars-v1.3.0.js
