+++ hbs-test/js/handlebars.runtime-v1.3.0.js
