+++ hbs-test/js/helpers/slugify.helper.js
@@ -1,0 +1,11 @@
+define(function() {
+	return function(text) {
+		return new Handlebars.SafeString(
+			text
+				.toLowerCase()
+				.replace(/ /g,'-')
+				.replace(/[^\w-]+/g,'')
+		);
+	}
+});
+
