+++ hbs-test/modules/tiny-module/cool.partial.hbs
@@ -1,0 +1,10 @@
+<style>
+.cool.partial {
+	background: rgba(0, 0, 0, .1);
+}
+</style>
+
+<div class="cool partial">
+	Coolest partial in town
+	{{! you can even have more partials in a partial }}
+</div>
