+++ hbs-test/modules/tiny-module/random-list-item.partial.hbs
@@ -1,0 +1,2 @@
+{{! We use a helper to make a slug url: }}
+<li><a href="{{@helpers/slugify.helper this}}">{{this}}</a></li></li>
