--- test/handlebars/handlebars.runtime.js
@@ -1,1 +1,0 @@
-../../handlebars.runtime.js
