--- test/hbs.js
@@ -1,1 +1,0 @@
-../hbs.js
