--- test/p.js
@@ -1,5 +1,0 @@
-define(function() {
-  return function(v) {
-    return '"' + v + '"';
-  }
-});
