--- test/test.hbs
@@ -1,1 +1,0 @@
-This is a {{@./p template}}
