--- test/test.html
@@ -1,19 +1,0 @@
-<!doctype html>
-  <script src="require.js"></script>
-  <script>
-    requirejs.config({
-      shim: {
-        handlebars: {
-          exports: 'Handlebars'
-        },
-        'handlebars/handlebars.runtime': {
-          exports: 'Handlebars'
-        }
-      }
-    });
-  </script>
-  <script>
-    require(['hbs!test'], function(test) {
-      document.body.innerHTML = test({template: 'handlebars template'})
-    });
-  </script>
